generator.py: Fix default value for conf member
SWUGenerator.conf is initialized as an empty dict, however it is
assigned the object returned by libconf.loads(), which is a AttrDict and
allows access to string keys via attribute access. However, Pylint
considers conf to be a dict and raises error E1101 when a conf attribute
is used. To be more clear in the code and also remove the Pylint error,
SWUGenerator.conf is intialized as an empty AttrDict.

Signed-off-by: Colin McAllister <colinmca242@gmail.com>