SWUGenerator: Fix close to cleanup temporary directory
There is no existing method to clean up the temporary directory for
SWUGenerator. This results in the temporary directory being left after
the program returns.

While this fixes the bug when close is called, it may be better to treat
SWUGenerator as a resource allocating class and add __enter__ and
__exit__ special methods such that SWUGenerator can be called with
"with" so __exit__ is always called and cleanup always occurs.

Signed-off-by: Colin McAllister <colinmca242@gmail.com>