refactor: remove setuptools from install_requires
Remove setuptools from install_requires as per PEP 518 [1]. Including
setuptools is mainly for documentation purposes, but it cannot satisfy
the requirement if it's not already satisfied (since setup.py imports
setuptools).

In this particular case, it's even worse, since setuptools is listed in
install_requires, but install_requires is used to specify the
dependencies required to run the package, while setup_requires is used
to specify the dependencies required to run the setup script itself.

[1] https://peps.python.org/pep-0518/

Signed-off-by: Michael Adler <michael.adler@siemens.com>
Acked-by: Stefano Babic <sbabic@denx.de>