Pass path for input and output for CUSTOM sign
Signing with custom tool is currently tricky because the path to the
file to be signed and to the siignature are not passed to the external
tool. Add automatically the path for sw-description and sw-description
as last two parameters before calling the script. The syntax will
become:

	--sign CUSTOM, <own tool>,<parms1>,..<parms-n>

And the generator will execute:

<own tool> <parms1> ..<parms-n> <path to sw-desc> <path to sig>

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>
Reported-by: Daniel Braunwarth <oss@braunwarth.dev>