Add module option to PKCS11 signing argument
The PKCS11 signing option uses pkcs11-tool for archive signing.
pkcs11-tool allows specifying the module to load using the --module
argument.

When SoftHSM is used, the module to load is specified via the input
argument.

This commit adds an optional input argument to specify the module to
load.The optional argument is added to the PKCS11 option alongside the PIN
argument. If no module is specified, the signing process will not include
the --module option.

Signed-off-by: Javier Fernandez Pastrana <javier.pastrana@linutronix.de>