fix: add missing `decrypted-size` entry
The decrypted-size field is required when flashing encrypted images
to the MTD partition on NOR flash.

This attribute is also required to handle the UBI Volumes. Due to a
limit in the Linux kernel API for UBI volumes, the size reserved to
be written on disk should be declared before actually writing anything.

Signed-off-by: Pratik Manvar <pratik.manvar@ifm.com>