fix: correct PKCS#11 mechanism usage in signing command
Fix an issue where the pkcs11-tool was invoked with an empty mechanism
due to a space after `-m ` flag. This caused signing to fail with the
error "Unknown PKCS11 mechanism".

So, remove space after `-m` to fix PKCS#11 mechanism error as below.
  ['-m ', 'SHA256-RSA-PKCS']
  ['-m', 'SHA256-RSA-PKCS']

The mechanism should now correctly set to `SHA256-RSA-PKCS`.

Signed-off-by: Pratik Manvar <pratik.manvar@ifm.com>