Revert "Handle multiple configuration files"
This reverts commit af6f21260b85c0c3c5c779d04d0ca5c6bc169c42.

After discussion on ML, this feature can be done in other way by using
the include directive of libconfig.

A config file could be created in this way:

variables : {
    @include "file1"
    @include "file2"
         .....
}

The commit breaks how the paramters are passed making them dependent of
the order in which the command line args are specified.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>
Reported-by: James Hilliard <james.hilliard1@gmail.com>
CC: Ernestas Kulik <ernestas.k@iconn-networks.com>